Переделал весь фрагмент Market, убрал лишнее в ProfileFragment

+++ app/src/main/java/com/example/testframe/ui/market/ItemMarket.kt
@@ -3,5 +3,5 @@ package com.example.testframe.ui.market
 import android.net.Uri
 
 data class ItemMarket(val title:String,
-                      val value:Long,
+                      val value:Int,
                       val uriToImg: Uri)
